+++ test-packages/test-services-e2e/v4/test-service/action-imports.js
@@ -1,6 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.actionImports = exports.createTestEntityReturnId = exports.createTestEntity = void 0;
+exports.actionImports = exports.createTestEntityByIdReturnId = exports.createTestEntityById = void 0;
 /*
  * Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved.
  *
@@ -9,33 +9,33 @@ exports.actionImports = exports.createTestEntityReturnId = exports.createTestEnt
 var core_1 = require("@sap-cloud-sdk/core");
 var TestEntity_1 = require("./TestEntity");
 /**
- * Create Test Entity.
+ * Create Test Entity By Id.
  *
  * @param parameters - Object containing all parameters for the action import.
  * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
  */
-function createTestEntity(parameters) {
+function createTestEntityById(parameters) {
     var params = {
         id: new core_1.ActionImportParameter('id', 'Edm.Int32', parameters.id)
     };
-    return new core_1.ActionImportRequestBuilder('/odata/test-service', 'createTestEntity', function (data) { return core_1.transformReturnValueForEntityV4(data, TestEntity_1.TestEntity); }, params);
+    return new core_1.ActionImportRequestBuilder('/odata/test-service', 'createTestEntityById', function (data) { return core_1.transformReturnValueForEntityV4(data, TestEntity_1.TestEntity); }, params);
 }
-exports.createTestEntity = createTestEntity;
+exports.createTestEntityById = createTestEntityById;
 /**
- * Create Test Entity Return Id.
+ * Create Test Entity By Id Return Id.
  *
  * @param parameters - Object containing all parameters for the action import.
  * @returns A request builder that allows to overwrite some of the values and execute the resulting request.
  */
-function createTestEntityReturnId(parameters) {
+function createTestEntityByIdReturnId(parameters) {
     var params = {
         id: new core_1.ActionImportParameter('id', 'Edm.Int32', parameters.id)
     };
-    return new core_1.ActionImportRequestBuilder('/odata/test-service', 'createTestEntityReturnId', function (data) { return core_1.transformReturnValueForEdmTypeV4(data, function (val) { return core_1.edmToTsV4(val.value, 'Edm.Int32'); }); }, params);
+    return new core_1.ActionImportRequestBuilder('/odata/test-service', 'createTestEntityByIdReturnId', function (data) { return core_1.transformReturnValueForEdmTypeV4(data, function (val) { return core_1.edmToTsV4(val.value, 'Edm.Int32'); }); }, params);
 }
-exports.createTestEntityReturnId = createTestEntityReturnId;
+exports.createTestEntityByIdReturnId = createTestEntityByIdReturnId;
 exports.actionImports = {
-    createTestEntity: createTestEntity,
-    createTestEntityReturnId: createTestEntityReturnId
+    createTestEntityById: createTestEntityById,
+    createTestEntityByIdReturnId: createTestEntityByIdReturnId
 };
 //# sourceMappingURL=action-imports.js.map