+++ test-packages/test-services/openapi/test-service/index.ts
@@ -9,3 +9,4 @@ export * from './test-case-api';
 export * from './tag-dot-api';
 export * from './tag-space-api';
 export * from './default-api';
+export * from './extension-api';